feat: add invalidate current user pipeline
